Added schedule page and styled home screen
