+++ ConverstationV01/app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#008577</color>
-    <color name="colorPrimaryDark">#00574B</color>
-    <color name="colorAccent">#D81B60</color>
+    <color name="colorPrimary">#0F9942</color>
+    <color name="colorPrimaryDark">#0036BE</color>
+    <color name="colorAccent">#488BF0</color>
 </resources>
