+++ ConverstationV01/app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">ConverstationV0.1</string>
     <string name="action_settings">Settings</string>
+    <string name="title_activity_schedule">schedule</string>
 </resources>
